Merged in docs_fix_prepushhook_error (pull request #5876)
Fix docs erros that cause prePushHook to fail

* Fix docs erros that cause prePushHook to fail

Adds documentation for the print module's `transferParameter` configuration option, including a description, schema, and example.

Removes the `transferParameter` example from the main configuration example, as it is now documented separately.

This resolves documentation errors and improves clarity regarding the usage of `transferParameter`.


Approved-by: Jasper Elbeshausen