[DOCS] revert markdown table formatting
While formatted Markdown tables certainly provide an improved editing experience in IDEs with Markdown support, some IDEs render these tables as terrible chaos. This is the result of line wrapping which some IDEs have enabled by default. Since it is not possible to configure EditorConfig to set a line wrapping policy, the formatting is undone.

This change only affects the raw Markdown files and has no impacted on the final HTML product documentation rendered with mkdocs.